Backend/requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
numpy
python-multipart
python-jose
passlib[bcrypt]
bcrypt==3.2.2
email-validator
requests
python-dotenv